backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # Async PostgreSQL driver
psycopg2-binary==2.9.9  # Sync PostgreSQL (for migrations)
pgvector==0.2.4  # PostgreSQL vector extension support

# AWS
boto3==1.34.34

# Vector Database
pinecone-client==3.0.2

# AI/ML
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1
groq==0.4.1  # Groq API client

# XML Parsing
lxml==5.1.0
defusedxml==0.7.1  # Secure XML parsing

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1

# Text Processing
numpy==1.26.3

# Reports Generation & Sharing
reportlab==4.0.9  # PDF generation
PyPDF2==3.0.1  # PDF manipulation
python-docx==1.1.0  # Word document generation
aiosmtplib==3.0.1  # Async email sending
